refactor emulated ecc
