+++ relation/src/gadgets/ecc/emulated/short_weierstrass.rs
@@ -1,0 +1,7 @@
+// Copyright (c) 2022 Espresso Systems (espressosys.com)
+// This file is part of the Jellyfish library.
+
+// You should have received a copy of the MIT License
+// along with the Jellyfish library. If not, see <https://mit-license.org/>.
+
+//! Short Weierstrass curve point addition
